Custom conf parser to separate file
