+++ modules/conf.py
@@ -3,6 +3,7 @@ import yaml
 from modules.conf_models import *
 from pydantic import ValidationError
 
+
 class Conf:
     def __init__(self, logger):
         self.logger = logger
@@ -17,7 +18,7 @@ class Conf:
             with open(config_file, "r") as file:
                 try:
                     user_config = yaml.safe_load(file) or {}
-                    BaseConf.validate(user_config)
+                    BaseConf.model_validate(user_config)
                     self.logger.debug("Validated user config")
                     config = BaseConf(**user_config)
                     self.logger.info("Parsed configuration file.")
@@ -29,11 +30,10 @@ class Conf:
             config = BaseConf()
 
         return config
-    
+
     def dump_config_as_yaml(self, conf: BaseConf, skip_defaults: bool, output_file: str = "config.yaml.dump"):
         settings_dict = conf.dict(skip_defaults=skip_defaults)
         yaml_str = yaml.dump(settings_dict, indent=4)
-    
-        with open(output_file, 'w') as file:
-            file.write(yaml_str)
 
+        with open(output_file, "w") as file:
+            file.write(yaml_str)
